requirements.txt
streamlit==1.28.0
pandas==2.0.3
numpy==1.24.3
opencv-python==4.8.0.76
pytesseract==0.3.10
pillow==10.0.0
python-dotenv==1.0.0
scikit-learn==1.3.0
imutils==0.5.4
pdf2image==1.16.3
spacy==3.6.1